Improve article page initialization and reducer reusability
This commit removes the 'navigate' dependency from the StoreProvider and optimizes the DynamicModuleLoader component. It also adds initial page state to the articlesPageSchema. A new service, initArticlesPage, has been introduced which checks if the articlesPage state is already initialized before dispatching the initState and fetchArticlesList actions. This commit also ensures that reducers are not being unnecessarily re-added when already present, improving scalability and application performance.