Update Article imports and add circular dependency plugin
The ArticleType and ArticleSortField imports have been updated to use the relative path. A new dependency, circular-dependency-plugin, has been added to the package.json and package-lock.json files. Its configuration is added to the buildPlugins.ts, which allows for detecting circular dependencies in the project during the webpack build phase. The plugin ensures the build fails in case of any circular dependencies thus helping keep the project structure clean and manageable.