Refactor codebase to use '@' alias for import paths
This commit changes imports throughout the project so that they use the '@' alias instead of relative paths. This makes the code more readable and easier to manage as the project scales. Furthermore, some linting changes have been made to styling and a new script for updating import paths has been added.