Update import paths and refactor Drawer component
The import paths in the preview.ts file for storybook have been updated to use absolute paths. The Drawer component has been refactored to remove the AnimationProvider from the NotificationButton component, and instead incorporated directly into the Drawer component itself. This makes the usage of the Drawer component self-contained and improves encapsulation.